More rename cleanup
