+++ GVFS/GVFS.Common/Git/HashingStream.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Security.Cryptography;
 
-namespace GVFS.Common
+namespace GVFS.Common.Git
 {
     public class HashingStream : Stream
     {
